Fixed pen up pen down problem, attempted wrapping

+++ src/constants/DisplayConstants.java
@@ -32,7 +32,7 @@ public class DisplayConstants {
 	public static final double HISTORY_Y = 2 * BUFFER;
 	
 	public static final double VIEW_HEIGHT = 480;
-	public static final int VIEW_WIDTH = 650;
+	public static final double VIEW_WIDTH = 650;
 	
 	public static final double COM_WIDTH = 250;
 	public static final double COM_HEIGHT = HISTORY_HEIGHT - VAR_HEIGHT - BUFFER;
